refactor: Move Manchester OWL syntax parsing to RequestManager
